requirements.txt
anthropic>=0.40.0
Pillow>=10.0
numpy>=1.24
opencv-python>=4.8
